move feature valiadtor tests
