+++ Frontend.Tests/ValidatorTests/Features/FeaturesReasonValidatorTests.cs
@@ -4,7 +4,7 @@ using Frontend.Validators.Features;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Xunit;
 
-namespace Frontend.Tests.ValidatorTests
+namespace Frontend.Tests.ValidatorTests.Features
 {
     public class FeaturesReasonValidatorTests
     {
